+++ bsn/rollup/cmd/rollup-fpd/daemon/recover_proof.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"path/filepath"
 
-	rollupfpcfg "github.com/babylonlabs-io/finality-provider/bsn/rollup-finality-provider/config"
+	rollupfpcfg "github.com/babylonlabs-io/finality-provider/bsn/rollup/config"
 	clientctx "github.com/babylonlabs-io/finality-provider/finality-provider/cmd/fpd/clientctx"
 	fpdaemon "github.com/babylonlabs-io/finality-provider/finality-provider/cmd/fpd/daemon"
 	"github.com/babylonlabs-io/finality-provider/util"
